chapter_1/requirements.txt
python-dotenv
huggingface-hub
markdownify
smolagents
requests
duckduckgo-search
pandas 